requirements.txt
# Core Python
python-dateutil==2.8.2
typing-extensions==4.8.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# UI Framework
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2

# API Documentation
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Caching
redis==5.0.1
hiredis==2.2.3

# Data Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Monitoring & Logging
python-json-logger==2.0.7
structlog==23.2.0

# Environment Management
python-dotenv==1.0.0

# Task Queue (future implementation)
# celery==5.3.4
# redis==5.0.1

# Cloud Storage (future implementation)
# boto3==1.29.7

# Monitoring Integration (future implementation)
# prometheus-client==0.19.0